Add files via upload
You will need a text editor such as notepad. Go to line 184 of the code then go to discord in a private server with just you I suggest and make a webhook. After you have created the webhook, copy the webhook URL and paste it in the quotation marks on line 184. Once finished, you can rename to make it look like a gamer or something and maybe ask someone to test your new game. It should open a command prompt and then close for them and then it will send you the info on the discord server with the webhook.